requirements.txt
aiogram>=3.7.0,<4.0.0
asyncpg>=0.29.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
tzdata>=2024.1
